Master

+++ src/main/groovy/org/clyze/doop/CommandLineAnalysisPostProcessor.groovy
@@ -1,22 +1,22 @@
 package org.clyze.doop
 
 import groovy.util.logging.Log4j
+import java.nio.file.Files
 import org.apache.commons.io.FileUtils
 import org.apache.commons.io.FilenameUtils
 import org.clyze.analysis.AnalysisPostProcessor
 import org.clyze.doop.core.Doop
 import org.clyze.doop.core.DoopAnalysis
-
-import java.nio.file.Files
+import org.clyze.utils.OS
 
 @Log4j
 class CommandLineAnalysisPostProcessor implements AnalysisPostProcessor<DoopAnalysis> {
 
 	@Override
 	void process(DoopAnalysis analysis) {
-		if (!analysis.options.X_STOP_AT_FACTS.value && !analysis.options.X_STOP_AT_BASIC.value)
+		if (!analysis.options.get('X_STOP_AT_FACTS').value && !analysis.options.get('X_STOP_AT_BASIC').value)
 			printStats(analysis)
-		if (analysis.options.SANITY.value && !analysis.options.X_DRY_RUN.value)
+		if (analysis.options.get('SANITY').value && !analysis.options.get('X_DRY_RUN').value)
 			printSanityResults(analysis)
 		linkResult(analysis)
 	}
@@ -29,7 +29,7 @@ class CommandLineAnalysisPostProcessor implements AnalysisPostProcessor<DoopAnal
 				if (!filterOutLBWarn(line)) lines << line
 			}
 		} else {
-			def file = new File("${analysis.database}/Stats_Runtime.csv")
+			def file = new File(analysis.database, 'Stats_Runtime.csv')
 			file.eachLine { String line -> lines << line.replace("\t", ", ") }
 		}
 
@@ -38,7 +38,7 @@ class CommandLineAnalysisPostProcessor implements AnalysisPostProcessor<DoopAnal
 			printf("%-80s %,d\n", it[0], it[1] as long)
 		}
 
-		if (!analysis.options.X_STATS_NONE.value && !analysis.options.X_DRY_RUN.value) {
+		if (!analysis.options.get('X_STATS_NONE').value && !analysis.options.get('X_DRY_RUN').value) {
 			lines = []
 
 			if (analysis.options.LB3.value) {
@@ -66,31 +66,37 @@ class CommandLineAnalysisPostProcessor implements AnalysisPostProcessor<DoopAnal
 	}
 
 	void linkResult(DoopAnalysis analysis) {
-		if (analysis.options.X_STOP_AT_FACTS.value) {
-			log.info "Making facts available at ${analysis.options.X_STOP_AT_FACTS.value}"
+		def xStopAtFactsOpt = analysis.options.get('X_STOP_AT_FACTS').value
+		if (xStopAtFactsOpt) {
+			log.info "Making facts available at ${xStopAtFactsOpt}"
 			return
 		}
 
 		def inputName
-		def platform = analysis.options.PLATFORM.value
+		def platform = analysis.options.get('PLATFORM').value
 
-		if (analysis.options.X_START_AFTER_FACTS.value)
+		if (analysis.options.get('X_START_AFTER_FACTS').value)
 			inputName = analysis.id
 		else
 			inputName = FilenameUtils.getBaseName(analysis.inputFiles[0].toString())
 
-		def humanDatabase = new File("${Doop.doopHome}/results/${inputName}/${analysis.name}/${platform}/${analysis.id}")
-		humanDatabase.mkdirs()
-		if (humanDatabase.exists()) {
-			FileUtils.deleteDirectory(humanDatabase)
-		}
-		log.info "Making database available at $humanDatabase"
-		Files.createSymbolicLink(humanDatabase.toPath(), analysis.database.toPath())
+		// Skip symbolic links on Windows.
+		if (OS.win) {
+			log.info "Making database available at $analysis.database"
+		} else {
+			def humanDatabase = new File("${Doop.doopHome}/results/${inputName}/${analysis.name}/${platform}/${analysis.id}")
+			humanDatabase.mkdirs()
+			if (humanDatabase.exists()) {
+				FileUtils.deleteDirectory(humanDatabase)
+			}
+			log.info "Making database available at $humanDatabase"
+			Files.createSymbolicLink(humanDatabase.toPath(), analysis.database.toPath())
 
-		def lastAnalysis = new File("${Doop.doopHome}/last-analysis")
-		Files.deleteIfExists(lastAnalysis.toPath())
-		log.info "Making database available at $lastAnalysis"
-		Files.createSymbolicLink(lastAnalysis.toPath(), analysis.database.toPath())
+			def lastAnalysis = new File("${Doop.doopHome}/last-analysis")
+			Files.deleteIfExists(lastAnalysis.toPath())
+			log.info "Making database available at $lastAnalysis"
+			Files.createSymbolicLink(lastAnalysis.toPath(), analysis.database.toPath())
+		}
 
 		if (analysis.options.SOUFFLE_PROFILE.value)
 			log.info "Souffle analysis profile available at ${analysis.outDir}/profile.txt"
